Handled lowercase problems and input without punctuation
It still responds very silly sometimes because the mirrored words are so limited. If I add more, they just become more complex and I'm not capable of doing that in this moment. I was also thinking of writing a test, but then I realized it needs an input so the test runs forever waiting for input. It doesn't work so I deleted it.